+++ src/test/java/automationpractice/com/features/checkout/PaymentFeatures.java
@@ -1,6 +1,5 @@
-package automationpractice.com.features.checkoutFeatures;
+package automationpractice.com.features.checkout;
 
-import automationpractice.com.features.AddressFeatures;
 import automationpractice.com.pages.*;
 import automationpractice.com.steps.LoginSteps;
 import cucumber.api.java.en.Then;
@@ -8,18 +7,13 @@ import cucumber.api.java.en.When;
 import net.thucydides.core.annotations.Steps;
 import org.junit.Assert;
 
-public class Payment {
+public class PaymentFeatures {
 
     private HomePage homePage;
-    private HeaderBox headerBox;
-    private LoginPage loginPage;
-    private MyAccountPage myAccountPage;
     private FirstProductPage firstProductPage;
     private CheckoutPage checkoutPage;
-    private AddressFeatures addressFeatures;
     @Steps
     private LoginSteps loginSteps;
-
 
     @When("^user select first item from homepage and added it into cart$")
     public void userSelectFirstItemFromHomepageAndAddedItIntoCart() {
@@ -28,28 +22,11 @@ public class Payment {
         firstProductPage.clickProceedToCheckoutButton();
     }
 
-    @When("^he confirms every step to make order and pay by bank wire$")
-    public void heConfirmsEveryStepToMakeOrderAndPayByBankWire() {
-        checkoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
-        checkoutPage.clickOnProceedToCheckoutButtonInAddressLabel();
-        checkoutPage.clickAgreeCheckbox();
-        checkoutPage.clickOnProceedToCheckoutButtonInShippingLabel();
-        checkoutPage.clickPayByBankWireButton();
-    }
-
     @Then("^payment by wire is made$")
     public void paymentByWireIsMade() {
         Assert.assertTrue(checkoutPage.isSelectedPayByBankWire());
     }
 
-    @When("^he confirms every step to make order and pay by check$")
-    public void heConfirmsEveryStepToMakeOrderAndPayByCheck() {
-        checkoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
-        checkoutPage.clickOnProceedToCheckoutButtonInAddressLabel();
-        checkoutPage.clickAgreeCheckbox();
-        checkoutPage.clickOnProceedToCheckoutButtonInShippingLabel();
-        checkoutPage.clickPayByCheckButton();
-    }
 
     @Then("^payment by check is made$")
     public void paymentByCheckIsMade() {
